+++ lib/dynamic_scaffold.rb
@@ -5,6 +5,7 @@ require 'dynamic_scaffold/controller_utilities'
 require 'dynamic_scaffold/controller'
 require 'dynamic_scaffold/config'
 require 'dynamic_scaffold/form_builder'
+require 'dynamic_scaffold/title'
 
 module DynamicScaffold
   module Error
